Se volvio a la idea del servidor, funciona con errores, se corregiran despues: la app se cierra porque el tiempo de espera es muy largo, aun asi se envia el mensaje y llega la notificación. se agregaran varios inputs de texto en donde se van a incluir: uno para la ip del servidor, otro para el titulo de la notificación, y otro para el mensaje de esta, y ahora el token se va a guardar en shared preferences para recuperarlo y que se pueda enviar todo desde un botón, tambien se va a considerar que cualquier dispositivo con la app les pueda eniar mensajes a otros, para ellos, se almacenarian los tokens en el servidor para enviar notificaciones a cada uno de estos dispositivos
