+++ Servidor/.gitignore
@@ -1,5 +1,5 @@
 # Ignore las claves privadas y los archivos de configuración
-Clave.json
+
 config/
 
 # Ignore los módulos instalados por npm
