+++ Servidor/package.json
@@ -1,7 +1,22 @@
 {
+  "name": "mi-servidor-mensajes",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "run": "node index.js",
+    "dev": "nodemon index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "hugo",
+  "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "cors": "^2.8.5",
     "express": "^4.18.2",
     "firebase-admin": "^11.7.0"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.22"
   }
 }
